+++ test/extensions/filters/network/mongo_proxy/proxy_test.cc
@@ -9,7 +9,9 @@
 #include "extensions/filters/network/mongo_proxy/bson_impl.h"
 #include "extensions/filters/network/mongo_proxy/codec_impl.h"
 #include "extensions/filters/network/mongo_proxy/proxy.h"
+#include "extensions/filters/network/well_known_names.h"
 
+#include "test/common/stream_info/test_util.h"
 #include "test/mocks/access_log/mocks.h"
 #include "test/mocks/event/mocks.h"
 #include "test/mocks/filesystem/mocks.h"
@@ -27,6 +29,7 @@ using testing::Invoke;
 using testing::NiceMock;
 using testing::Property;
 using testing::Return;
+using testing::ReturnRef;
 
 namespace Envoy {
 namespace Extensions {
@@ -69,14 +72,19 @@ public:
     ON_CALL(runtime_.snapshot_, featureEnabled("mongo.logging_enabled", 100))
         .WillByDefault(Return(true));
 
+    EXPECT_CALL(read_filter_callbacks_, connection())
+        .WillRepeatedly(ReturnRef(read_filter_callbacks_.connection_));
+    EXPECT_CALL(read_filter_callbacks_.connection_, streamInfo())
+        .WillRepeatedly(ReturnRef(stream_info_));
+
     EXPECT_CALL(log_manager_, createAccessLog(_)).WillOnce(Return(file_));
     access_log_.reset(new AccessLog("test", log_manager_, dispatcher_.timeSystem()));
   }
 
   void initializeFilter() {
-    filter_ =
-        std::make_unique<TestProxyFilter>("test.", store_, runtime_, access_log_, fault_config_,
-                                          drain_decision_, generator_, dispatcher_.timeSystem());
+    filter_ = std::make_unique<TestProxyFilter>("test.", store_, runtime_, access_log_,
+                                                fault_config_, drain_decision_, generator_,
+                                                dispatcher_.timeSystem(), true);
     filter_->initializeReadFilterCallbacks(read_filter_callbacks_);
     filter_->onNewConnection();
 
@@ -115,6 +123,7 @@ public:
   Envoy::AccessLog::MockAccessLogManager log_manager_;
   NiceMock<Network::MockDrainDecision> drain_decision_;
   NiceMock<Runtime::MockRandomGenerator> generator_;
+  TestStreamInfo stream_info_;
 };
 
 TEST_F(MongoProxyFilterTest, DelayFaults) {
@@ -189,6 +198,104 @@ TEST_F(MongoProxyFilterTest, DelayFaultsRuntimeDisabled) {
 
   EXPECT_EQ(Network::FilterStatus::Continue, filter_->onData(fake_data_, false));
   EXPECT_EQ(0U, store_.counter("test.delays_injected").value());
+}
+
+TEST_F(MongoProxyFilterTest, DynamicMetadata) {
+  initializeFilter();
+
+  EXPECT_CALL(*file_, write(_)).Times(AtLeast(1));
+
+  EXPECT_CALL(*filter_->decoder_, onData(_)).WillOnce(Invoke([&](Buffer::Instance&) -> void {
+    QueryMessagePtr message(new QueryMessageImpl(0, 0));
+    message->fullCollectionName("db.test");
+    message->flags(0b1110010);
+    message->query(Bson::DocumentImpl::create());
+    filter_->callbacks_->decodeQuery(std::move(message));
+  }));
+  filter_->onData(fake_data_, false);
+
+  auto& metadata =
+      stream_info_.dynamicMetadata().filter_metadata().at(NetworkFilterNames::get().MongoProxy);
+  auto& query_message = metadata.fields().at("messages").list_value().values(0).struct_value();
+  EXPECT_EQ("OP_QUERY", query_message.fields().at("operation").string_value());
+  EXPECT_EQ("db", query_message.fields().at("database").string_value());
+  EXPECT_EQ("test", query_message.fields().at("collection").string_value());
+
+  EXPECT_CALL(*filter_->decoder_, onData(_)).WillOnce(Invoke([&](Buffer::Instance&) -> void {
+    ReplyMessagePtr message(new ReplyMessageImpl(0, 0));
+    message->flags(0b11);
+    message->cursorId(1);
+    message->documents().push_back(Bson::DocumentImpl::create()->addString("hello", "world"));
+    filter_->callbacks_->decodeReply(std::move(message));
+  }));
+  filter_->onWrite(fake_data_, false);
+
+  auto& reply_message = metadata.fields().at("messages").list_value().values(0).struct_value();
+  EXPECT_EQ("OP_REPLY", reply_message.fields().at("operation").string_value());
+
+  EXPECT_CALL(*filter_->decoder_, onData(_)).WillOnce(Invoke([&](Buffer::Instance&) -> void {
+    GetMoreMessagePtr message(new GetMoreMessageImpl(0, 0));
+    message->fullCollectionName("db.test");
+    message->cursorId(1);
+    filter_->callbacks_->decodeGetMore(std::move(message));
+  }));
+  filter_->onData(fake_data_, false);
+
+  auto& get_more_message = metadata.fields().at("messages").list_value().values(0).struct_value();
+  EXPECT_EQ("OP_GET_MORE", get_more_message.fields().at("operation").string_value());
+  EXPECT_EQ("db", get_more_message.fields().at("database").string_value());
+  EXPECT_EQ("test", get_more_message.fields().at("collection").string_value());
+
+  EXPECT_CALL(*filter_->decoder_, onData(_)).WillOnce(Invoke([&](Buffer::Instance&) -> void {
+    InsertMessagePtr message(new InsertMessageImpl(0, 0));
+    message->fullCollectionName("db.test");
+    message->documents().push_back(Bson::DocumentImpl::create());
+    filter_->callbacks_->decodeInsert(std::move(message));
+  }));
+  filter_->onData(fake_data_, false);
+
+  auto& insert_message = metadata.fields().at("messages").list_value().values(0).struct_value();
+  EXPECT_EQ("OP_INSERT", insert_message.fields().at("operation").string_value());
+  EXPECT_EQ("db", insert_message.fields().at("database").string_value());
+  EXPECT_EQ("test", insert_message.fields().at("collection").string_value());
+
+  EXPECT_CALL(*filter_->decoder_, onData(_)).WillOnce(Invoke([&](Buffer::Instance&) -> void {
+    KillCursorsMessagePtr message(new KillCursorsMessageImpl(0, 0));
+    message->numberOfCursorIds(1);
+    message->cursorIds({1});
+    filter_->callbacks_->decodeKillCursors(std::move(message));
+  }));
+  filter_->onData(fake_data_, false);
+
+  auto& kill_cursors_message =
+      metadata.fields().at("messages").list_value().values(0).struct_value();
+  EXPECT_EQ("OP_KILL_CURSORS", kill_cursors_message.fields().at("operation").string_value());
+
+  EXPECT_CALL(*filter_->decoder_, onData(_)).WillOnce(Invoke([&](Buffer::Instance&) -> void {
+    CommandMessagePtr message(new CommandMessageImpl(0, 0));
+    message->database(std::string("db"));
+    message->commandName(std::string("command"));
+    message->metadata(Bson::DocumentImpl::create());
+    message->commandArgs(Bson::DocumentImpl::create());
+    filter_->callbacks_->decodeCommand(std::move(message));
+  }));
+  filter_->onData(fake_data_, false);
+
+  auto& command_message = metadata.fields().at("messages").list_value().values(0).struct_value();
+  EXPECT_EQ("OP_COMMAND", command_message.fields().at("operation").string_value());
+  EXPECT_EQ("db", command_message.fields().at("database").string_value());
+
+  EXPECT_CALL(*filter_->decoder_, onData(_)).WillOnce(Invoke([&](Buffer::Instance&) -> void {
+    CommandReplyMessagePtr message(new CommandReplyMessageImpl(0, 0));
+    message->metadata(Bson::DocumentImpl::create());
+    message->commandReply(Bson::DocumentImpl::create());
+    filter_->callbacks_->decodeCommandReply(std::move(message));
+  }));
+  filter_->onData(fake_data_, false);
+
+  auto& command_reply_message =
+      metadata.fields().at("messages").list_value().values(0).struct_value();
+  EXPECT_EQ("OP_COMMANDREPLY", command_reply_message.fields().at("operation").string_value());
 }
 
 TEST_F(MongoProxyFilterTest, Stats) {